requirements.txt
asciimatics==1.13.0
attrs==21.4.0
future==0.18.2
iniconfig==1.1.1
packaging==21.3
Pillow==9.0.1
pluggy==1.0.0
py==1.11.0
pyfiglet==0.8.post1
pyparsing==3.0.7
pyTermTk==0.9.0a37
pytest==7.1.1
single-source==0.2.0
strip-ansi==0.1.1
textual==0.1.17
toml==0.10.2
wcwidth==0.2.5
